+++ integrations/axum/src/lib.rs
@@ -4,9 +4,11 @@
 #![warn(missing_docs)]
 
 mod extract;
+mod query;
 mod response;
 mod subscription;
 
 pub use extract::{GraphQLBatchRequest, GraphQLRequest};
+pub use query::GraphQL;
 pub use response::GraphQLResponse;
 pub use subscription::{GraphQLProtocol, GraphQLSubscription, GraphQLWebSocket};